+++ src/main/java/api/courier/Courier.java
@@ -1,30 +1,32 @@
-public class Credentials {
+package api.courier;
 
+public class Courier {
     private String login;
     private String password;
+    private String firstName;
 
-    public Credentials(String login, String password) {
+    public Courier(String login, String password, String firstName) {
         this.login = login;
         this.password = password;
+        this.firstName = firstName;
     }
 
     public String getLogin() {
         return login;
     }
-
     public void setLogin(String login) {
         this.login = login;
     }
-
     public String getPassword() {
         return password;
     }
-
     public void setPassword(String password) {
         this.password = password;
     }
-
-    public static Credentials from(Courier courier) {
-        return new Credentials(courier.getLogin(), courier.getPassword());
+    public String getFirstName() {
+        return firstName;
+    }
+    public void setFirstName(String firstName) {
+        this.firstName = firstName;
     }
 }
