+++ src/main/java/api/courier/Credentials.java
@@ -1,12 +1,13 @@
-public class Courier {
+package api.courier;
+
+public class Credentials {
+
     private String login;
     private String password;
-    private String firstName;
 
-    public Courier(String login, String password, String firstName) {
+    public Credentials(String login, String password) {
         this.login = login;
         this.password = password;
-        this.firstName = firstName;
     }
 
     public String getLogin() {
@@ -25,13 +26,7 @@ public class Courier {
         this.password = password;
     }
 
-    public String getFirstName() {
-        return firstName;
+    public static Credentials from(Courier courier) {
+        return new Credentials(courier.getLogin(), courier.getPassword());
     }
-
-    public void setFirstName(String firstName) {
-        this.firstName = firstName;
-    }
-
-
 }
