+++ src/main/java/api/utils/RestBase.java
@@ -1,3 +1,5 @@
+package api.utils;
+
 import io.restassured.builder.RequestSpecBuilder;
 import io.restassured.http.ContentType;
 import io.restassured.specification.RequestSpecification;
@@ -11,6 +13,5 @@ public class RestBase {
                 .setBaseUri(BASE_URL)
                 .setContentType(ContentType.JSON)
                 .build();
-
     }
 }
