+++ src/test/java/CourierCreationTest.java
@@ -1,3 +1,7 @@
+import api.courier.Courier;
+import api.courier.CourierClient;
+import api.courier.CourierGenerator;
+import api.courier.Credentials;
 import io.qameta.allure.Description;
 import io.qameta.allure.junit4.DisplayName;
 import io.restassured.response.ValidatableResponse;
@@ -28,25 +32,23 @@ public class CourierCreationTest {
     }
 
     @Test
-    @DisplayName("Courier can be created successfully")
+    @DisplayName("api.courier.Courier can be created successfully")
     @Description("Service will return 201 Created when new courier is created")
     public void courierCanBeCreatedTest() {
-        //Создание и проверка, что курьер был создан
+        //Создание курьера
         ValidatableResponse response = courierClient.create(courier);
+        //Авторизация курьера
+        ValidatableResponse loginResponse = courierClient.login(Credentials.from(courier));
+        // Проверка успешного создания курьера
         int statusCode = response.extract().statusCode();
         assertEquals("Status code is incorrect", SC_CREATED, statusCode);
         boolean isCreated = response.extract().path("ok");
-        assertTrue("Courier is not created", isCreated);
-        //Авторизация курьера
-        ValidatableResponse loginResponse = courierClient.login(Credentials.from(courier));
+        assertTrue("api.courier.Courier is not created", isCreated);
+        //Проверка авторизации курьера
         int loginStatusCode = loginResponse.extract().statusCode();
-        //Проверка успешной авторизации
-        assertEquals("Courier is not login", SC_OK, loginStatusCode);
+        assertEquals("api.courier.Courier is not login", SC_OK, loginStatusCode);
         courierId = loginResponse.extract().path("id");
         //Проверка успешного запроса на возврат id
         assertNotNull("Id is null", courierId);
-
     }
-
-
 }
