+++ src/test/java/CourierCreationWithoutPassTest.java
@@ -1,3 +1,6 @@
+import api.courier.Courier;
+import api.courier.CourierClient;
+import api.courier.CourierGenerator;
 import io.qameta.allure.Description;
 import io.qameta.allure.junit4.DisplayName;
 import io.restassured.response.ValidatableResponse;
@@ -19,7 +22,7 @@ public class CourierCreationWithoutPassTest {
     }
 
     @Test
-    @DisplayName("Courier creation without password")
+    @DisplayName("api.courier.Courier creation without password")
     @Description("Service will return 400 Bad Request if courier creation will be without password")
     public void courierCannotBeCreatedTest () {
         //Создание и проверка, что процесс не удался
@@ -27,5 +30,4 @@ public class CourierCreationWithoutPassTest {
         int statusCode = response.extract().statusCode();
         assertEquals("Status code is incorrect", SC_BAD_REQUEST, statusCode);
     }
-
 }
