+++ src/test/java/DoubleCourierCreationTest.java
@@ -1,3 +1,7 @@
+import api.courier.Courier;
+import api.courier.CourierClient;
+import api.courier.CourierGenerator;
+import api.courier.Credentials;
 import io.qameta.allure.Description;
 import io.qameta.allure.junit4.DisplayName;
 import io.restassured.response.ValidatableResponse;
@@ -18,7 +22,6 @@ public class DoubleCourierCreationTest {
     public void setUp() {
         courier = CourierGenerator.courierDefault();
         courierClient = new CourierClient();
-
     }
 
     @After
@@ -30,21 +33,25 @@ public class DoubleCourierCreationTest {
     @DisplayName("Client can't create double courier")
     @Description("Service will return 400 Bad Request when double courier is created")
     public void doubleCourierCreationTest() {
-        //Создание курьера и проверка успеха
+        //Создание курьера
         ValidatableResponse response = courierClient.create(courier);
+        //Логин
+        ValidatableResponse loginResponse = courierClient.login(Credentials.from(courier));
+        //Повторное создание курьера
+        ValidatableResponse doubleResponse = courierClient.create(courier);
+
+        // Проверка успешного создания курьера
         int statusCode = response.extract().statusCode();
         assertEquals("Status code is incorrect", SC_CREATED, statusCode);
         boolean isCreated = response.extract().path("ok");
-        assertTrue("Courier is not created", isCreated);
-        //Логин для получения id
-        ValidatableResponse loginResponse = courierClient.login(Credentials.from(courier));
+        assertTrue("api.courier.Courier is not created", isCreated);
+        //Проверка логина и получение id
         int loginStatusCode = loginResponse.extract().statusCode();
-        assertEquals("Courier is not login", SC_OK, loginStatusCode);
+        assertEquals("api.courier.Courier is not login", SC_OK, loginStatusCode);
         courierId = loginResponse.extract().path("id");
         assertNotNull("Id is null", courierId);
-        //Повторное создание курьера
-        ValidatableResponse doubleResponse = courierClient.create(courier);
+        //Проверка, что повторное создание того же курьера вызывает ошибку
         int doubleStatusCode = doubleResponse.extract().statusCode();
-        assertEquals("Status code is incorrect",SC_CONFLICT, doubleStatusCode);
+        assertEquals("Status code is not 400",SC_CONFLICT, doubleStatusCode);
     }
 }
